+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/resttemplate/EnhancedRestTemplateReporter.java
@@ -17,12 +17,12 @@
 
 package com.tencent.cloud.rpc.enhancement.resttemplate;
 
-import java.net.HttpURLConnection;
+import java.io.IOException;
 import java.net.URI;
-import java.net.URL;
+import java.util.Map;
 
 import com.tencent.cloud.common.metadata.MetadataContext;
-import com.tencent.cloud.common.util.ReflectionUtils;
+import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.rpc.enhancement.AbstractPolarisReporterAdapter;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
 import com.tencent.polaris.api.core.ConsumerAPI;
@@ -33,6 +33,9 @@ import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.springframework.beans.BeansException;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.lang.NonNull;
@@ -43,13 +46,14 @@ import org.springframework.web.client.ResponseErrorHandler;
  *
  * @author wh 2022/6/21
  */
-public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter implements ResponseErrorHandler {
+public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter implements ResponseErrorHandler, ApplicationContextAware {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(EnhancedRestTemplateReporter.class);
 
-	private static final String FIELD_NAME = "connection";
+	static final String HEADER_HAS_ERROR = "X-SCT-Has-Error";
 
 	private final ConsumerAPI consumerAPI;
+	private ResponseErrorHandler delegateHandler;
 
 	public EnhancedRestTemplateReporter(RpcEnhancementReporterProperties properties, ConsumerAPI consumerAPI) {
 		super(properties);
@@ -57,29 +61,68 @@ public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter
 	}
 
 	@Override
-	public boolean hasError(@NonNull ClientHttpResponse response) {
+	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+		String[] handlerBeanNames = applicationContext.getBeanNamesForType(ResponseErrorHandler.class);
+		if (handlerBeanNames.length == 1) {
+			return;
+		}
+
+		// inject user custom ResponseErrorHandler
+		for (String beanName : handlerBeanNames) {
+			// ignore self
+			if (StringUtils.equalsIgnoreCase("enhancedRestTemplateReporter", beanName)) {
+				continue;
+			}
+			this.delegateHandler = (ResponseErrorHandler) applicationContext.getBean(beanName);
+		}
+	}
+
+	@Override
+	public boolean hasError(@NonNull ClientHttpResponse response) throws IOException {
+		if (delegateHandler != null) {
+			// Preserve the delegated handler result
+			boolean hasError = delegateHandler.hasError(response);
+			response.getHeaders().add(HEADER_HAS_ERROR, String.valueOf(hasError));
+		}
 		return true;
 	}
 
 	@Override
-	public void handleError(@NonNull ClientHttpResponse response) {
+	public void handleError(@NonNull ClientHttpResponse response) throws IOException {
+		if (realHasError(response)) {
+			delegateHandler.handleError(response);
+		}
+
+		clear(response);
 	}
 
 	@Override
-	public void handleError(@NonNull URI url, @NonNull HttpMethod method, @NonNull ClientHttpResponse response) {
-		if (!properties.isEnabled()) {
-			return;
+	public void handleError(@NonNull URI url, @NonNull HttpMethod method, @NonNull ClientHttpResponse response) throws IOException {
+		// report result to polaris
+		if (reportProperties.isEnabled()) {
+			reportResult(url, response);
 		}
 
+		// invoke delegate handler
+		invokeDelegateHandler(url, method, response);
+	}
+
+	private void reportResult(URI url, ClientHttpResponse response) {
 		ServiceCallResult resultRequest = createServiceCallResult(url);
 		try {
+			Map<String, String> loadBalancerContext = MetadataContextHolder.get()
+					.getFragmentContext(MetadataContext.FRAGMENT_LOAD_BALANCER);
 
-			HttpURLConnection connection = (HttpURLConnection) ReflectionUtils.getFieldValue(response, FIELD_NAME);
-			if (connection != null) {
-				URL realURL = connection.getURL();
-				resultRequest.setHost(realURL.getHost());
-				resultRequest.setPort(realURL.getPort());
+			String targetHost = loadBalancerContext.get("host");
+			String targetPort = loadBalancerContext.get("port");
+
+			if (StringUtils.isBlank(targetHost) || StringUtils.isBlank(targetPort)) {
+				LOGGER.warn("Can not get target host or port from metadata context. host = {}, port = {}", targetHost, targetPort);
+				return;
 			}
+
+			resultRequest.setHost(targetHost);
+			resultRequest.setPort(Integer.parseInt(targetPort));
 
 			// checking response http status code
 			if (apply(response.getStatusCode())) {
@@ -96,6 +139,34 @@ public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter
 		}
 	}
 
+	private void invokeDelegateHandler(URI url, HttpMethod method, ClientHttpResponse response) throws IOException {
+		if (realHasError(response)) {
+			delegateHandler.handleError(url, method, response);
+		}
+
+		clear(response);
+	}
+
+	private Boolean realHasError(ClientHttpResponse response) {
+		if (delegateHandler == null) {
+			return false;
+		}
+
+		String hasErrorHeader = response.getHeaders().getFirst(HEADER_HAS_ERROR);
+		if (StringUtils.isBlank(hasErrorHeader)) {
+			return false;
+		}
+
+		return Boolean.parseBoolean(hasErrorHeader);
+	}
+
+	private void clear(ClientHttpResponse response) {
+		if (!response.getHeaders().containsKey(HEADER_HAS_ERROR)) {
+			return;
+		}
+		response.getHeaders().remove(HEADER_HAS_ERROR);
+	}
+
 	private ServiceCallResult createServiceCallResult(URI uri) {
 		ServiceCallResult resultRequest = new ServiceCallResult();
 		String serviceName = uri.getHost();
@@ -109,5 +180,9 @@ public class EnhancedRestTemplateReporter extends AbstractPolarisReporterAdapter
 			resultRequest.setCallerService(new ServiceKey(sourceNamespace, sourceService));
 		}
 		return resultRequest;
+	}
+
+	public void setDelegateHandler(ResponseErrorHandler delegateHandler) {
+		this.delegateHandler = delegateHandler;
 	}
 }